7-Zip Version 15.05

+++ C/Lzma2Enc.c
@@ -238,9 +238,9 @@ void Lzma2EncProps_Normalize(CLzma2EncProps *p)
     if (temp > p->lzmaProps.reduceSize)
     {
       UInt64 numBlocks = temp / p->blockSize;
-      if (numBlocks < t2)
+      if (numBlocks < (unsigned)t2)
       {
-        t2 = (UInt32)numBlocks;
+        t2 = (unsigned)numBlocks;
         t3 = t1 * t2;
       }
     }
